Moved THRFormsCcsRecordAggregationHelper to own file
